Chap03の書き直しで不要なソースコードの削除。

+++ Chap03/src/DivisionOperation2.java
@@ -3,7 +3,7 @@ public class DivisionOperation2 {
 		int num1 = 6;
 		int num2 = 5;
 
-		double ans = num1 / num2;
+		double ans = (double)num1 / num2;
 
 		System.out.println(ans);
 	}
